Stashing the changes.
