+++ oss_src/unity/server/unity_server_init.cpp
@@ -5,7 +5,6 @@
  * This software may be modified and distributed under the terms
  * of the BSD license. See the LICENSE file for details.
  */
-#include <unity/extensions/cumulative_aggregates.hpp>
 #include <unity/lib/simple_model.hpp>
 #include <unity/lib/toolkit_class_registry.hpp>
 #include <unity/lib/toolkit_function_registry.hpp>
@@ -16,7 +15,6 @@
 graphlab::toolkit_function_registry* init_toolkits() {
   graphlab::toolkit_function_registry* g_toolkit_functions = new graphlab::toolkit_function_registry();
   g_toolkit_functions->register_toolkit_function(graphlab::image_util::get_toolkit_function_registration());
-  g_toolkit_functions->register_toolkit_function(graphlab::cumulative_aggregates::get_toolkit_function_registration());
   return g_toolkit_functions;
 }
